% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreAlg.R
\name{CoreAlg}
\alias{CoreAlg}
\title{Core algorithm}
\usage{
CoreAlg(X, y)
}
\arguments{
\item{X}{cell-specific gene expression}

\item{y}{mixed expression per sample}
}
\description{
The core algorithm of CIBERSORT which is used svm
}
