% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doPerm.R
\name{doPerm}
\alias{doPerm}
\title{do permutations}
\usage{
doPerm(perm, X, Y)
}
\arguments{
\item{perm}{Number of permutations}

\item{X}{cell-specific gene expression}

\item{Y}{mixed expression per sample}
}
\description{
Do the permutations analysis
}
