% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untils.R
\name{enableParallel}
\alias{enableParallel}
\title{open multiple threads process}
\usage{
enableParallel(nThreads = NULL, maxSize = 500, verbose = FALSE)
}
\arguments{
\item{nThreads}{The number of threads}

\item{maxSize}{The max memory size in global, default 500MB, the unit is MB}

\item{verbose}{output other useful information}
}
\description{
open multiple threads process
}
\examples{
\dontrun{
  enableParallel()
}
}
