% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LM22}
\alias{LM22}
\title{The gene expression matrix from isolated 22 cells}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 547 rows and 22 columns.
}
\usage{
LM22
}
\description{
The gene expression matrix from isolated 22 cells
}
\keyword{datasets}
