% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mixed_expr}
\alias{mixed_expr}
\title{The heterogenous mixed expression matrix}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 20501 rows and 5 columns.
}
\usage{
mixed_expr
}
\description{
The heterogenous mixed expression matrix
}
\keyword{datasets}
