% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cibersort.R
\name{runCibersort}
\alias{runCibersort}
\title{Cibersort interface}
\usage{
runCibersort(
  mixtrix,
  sig_matrix = NULL,
  perm = 0,
  QN = TRUE,
  which_sig = "LM22"
)
}
\arguments{
\item{mixtrix}{heterogenous mixed expression matrix}

\item{sig_matrix}{file path to gene expression from isolated cells}

\item{perm}{Number of permutations}

\item{QN}{Perform quantile normalization or not (TRUE/FALSE)}

\item{which_sig}{select sig matrix. if sig_matrix is not NULL, this option will be ignored,default:LM22}
}
\description{
Cibersort interface
}
