% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIBERSORT-package.R, R/CIBERSORT.R
\docType{data}
\name{CIBERSORT}
\alias{CIBERSORT}
\title{Main functions}
\format{
An object of class \code{function} of length 1.
}
\usage{
CIBERSORT

CIBERSORT(sig_matrix, mixture_file, perm = 0, QN = TRUE)
}
\arguments{
\item{sig_matrix}{gene expression matrix from isolated cells.
row names are gene names, column names are cell names}

\item{mixture_file}{heterogenous mixed expression matrix.
row names are gene names, column names are sample names}

\item{perm}{Number of permutations}

\item{QN}{Perform quantile normalization or not (TRUE/FALSE)}
}
\description{
This is a package version of CIBERSORT

The Main function of CIBERSORT
}
\examples{
\dontrun{
  sig_matrix <- system.file("extdata", "LM22.txt", package = "CIBERSORT")
  mixture_file <- system.file("extdata", "exampleForLUAD.txt", package = "CIBERSORT")
  results <- CIBERSORT('sig_matrix_file.txt','mixture_file.txt', perm, QN)
}
}
\keyword{datasets}
