% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIBERSORT.R
\name{cibersort}
\alias{cibersort}
\title{Main functions}
\usage{
cibersort(sig_matrix, mixture_file, perm = 0, QN = TRUE)
}
\arguments{
\item{sig_matrix}{sig_matrix file path to gene expression from isolated cells}

\item{mixture_file}{mixture_file heterogenous mixed expression}

\item{perm}{Number of permutations}

\item{QN}{Perform quantile normalization or not (TRUE/FALSE)}
}
\description{
The Main function of CIBERSORT
}
\examples{
\dontrun{
  sig_matrix <- system.file("extdata", "LM22.txt", package = "CIBERSORT")
  mixture_file <- system.file("extdata", "exampleForLUAD.txt", package = "CIBERSORT")
  results <- cibersort(sig_matrix, mixture_file)
}
}
